Added test for API endpoint using JEST and supertest, moved routes to app.js from index.js
