+++ app.js
@@ -1,0 +1,13 @@
+const express = require('express')
+const app = express()
+app.use(express.json())
+
+app.get('/', (req, res) => {
+    res.send('<h1>Hello World!</h1>')
+})
+
+app.get('/api/ping/', (req, res) => {
+    const json = {value: 'pong'}
+    res.json(json)
+})
+module.exports = app
