+++ index.js
@@ -1,18 +1,8 @@
-const express = require('express')
-const app = express()
-
-app.use(express.json())
-
-app.get('/', (req, res) => {
-    res.send('<h1>Hello World!</h1>')
-})
-
-app.get('/api/ping/', (req, res) => {
-    const json = {value: 'pong'}
-    res.json(json)
-})
+const app = require('./app')
+const http = require('http')
+const server = http.createServer(app)
 
 const PORT = 3001
-app.listen(PORT, () => {
+server.listen(PORT, () => {
     console.log('hello world!')
 })
