+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node index.js",
     "dev": "nodemon index.js",
-    "test": "\"echo \\\"No test spesified\\\" && exit 1\"",
+    "test": "cross-env NODE_ENV=test jest --verbose --detectOpenHandles",
     "lint": "eslint .",
     "testlint": "eslint . --fix"
   },
@@ -25,10 +25,19 @@
   },
   "homepage": "https://github.com/hy-sql/hy-sql#readme",
   "dependencies": {
+    "cross-env": "^7.0.2",
     "express": "^4.17.1"
   },
   "devDependencies": {
     "eslint": "^7.0.0",
-    "nodemon": "^2.0.3"
+    "jest": "^26.0.1",
+    "nodemon": "^2.0.3",
+    "supertest": "^4.0.2"
+  },
+  "jest": {
+    "testEnvironment": "node",
+    "coveragePathIgnorePatterns": [
+      "/node_modules/"
+    ]
   }
 }
